Mark the indels that lies in the tandem repeat and lower their weight

+++ PhasingGraph.h
@@ -64,7 +64,7 @@ struct VariantEdge{
     // node pair 
     std::pair<PosAllele,PosAllele> findBestEdgePair(int targetPos, bool isONT, double diffRatioThreshold, bool debug, float &weight, VoteResult &vote);
     // number of read of two node. AA and AB combination
-    std::pair<int,int> findNumberOfRead(int targetPos);
+    std::pair<float,float> findNumberOfRead(int targetPos);
 };
